Update - 리뷰 도메인 개선 및 예외 처리 추가 - 별점 필드 타입 float → BigDecimal로 변경 - 별점 검증 및 0.5 단위 값만 허용하는 로직 추가 - Review 엔티티에 예약 ID 기반 유니크 제약 조건 설정 - Review에 경험 평가 추가 메서드 추가 및 테스트 케이스 수정 - ReviewRepository에 fetch join 메서드 추가 - Review 생성 로직에서 별점 검증 및 fetch join 재조회 포함

+++ src/main/java/coffeandcommit/crema/domain/review/dto/request/ReviewRequestDTO.java
@@ -7,6 +7,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -20,7 +21,7 @@ public class ReviewRequestDTO {
 
     @DecimalMin(value = "0.0")
     @DecimalMax(value = "5.0")
-    private float starReview; // 별점 (1~5)
+    private BigDecimal starReview; // 별점 (1~5)
 
     @NotBlank
     @Size(min = 10, max = 500)
